refactor: rename bridge module to gateway, drop bridged-token wire string
Tracks the matching flashnet-execution rename
(flashnetxyz/flashnet-execution#427). Public API breaks below; consumers
need to bump together with that PR.

- src/execution/bridge.ts -> src/execution/gateway.ts
- src/execution/abis/sparkBridge.ts -> src/execution/abis/sparkGateway.ts
- sparkBridgeAbi -> sparkGatewayAbi
- queryBridgedTokenAddress -> querySparkTokenAddress
  (param bridgeAddress -> gatewayAddress)
- waitForBridgedTokenAddress -> waitForSparkTokenAddress
- SPARK_BRIDGE_REVERT_ERRORS -> SPARK_GATEWAY_REVERT_ERRORS
- Wire string "BridgedToken" -> "SparkToken" (DepositAsset payload)
- Doc comments and examples updated
- Re-exports in index.ts updated

Btkn shorthand kept everywhere (Btkn = Bitcoin Token, a Spark concept,
distinct from "bridged token"). depositBtkn / withdrawBtkn / encodeWithdrawBtkn
unchanged.

Conductor.SparkBridgeNotSet() entry left untouched in the revert table —
that selector belongs to flashnet-conductor, not to this rename.

Type-check, build, and 65/65 jest tests pass.
