+++ index.ts
@@ -7,8 +7,8 @@ export {
   Conductor,
   encodeWithdrawSats,
   encodeWithdrawToken,
-  queryBridgedTokenAddress,
-  waitForBridgedTokenAddress,
+  querySparkTokenAddress,
+  waitForSparkTokenAddress,
   encodeCreateBTCPool,
   encodeCreatePoolParams,
   fetchTokenInfo,
@@ -48,7 +48,7 @@ export {
   decodeRevertReason,
   DEFAULT_REVERT_ERRORS,
   CONDUCTOR_REVERT_ERRORS,
-  SPARK_BRIDGE_REVERT_ERRORS,
+  SPARK_GATEWAY_REVERT_ERRORS,
   SOLIDITY_BUILTIN_REVERT_ERRORS,
   traceInnermostRevert,
   extractTxHashFromStatusMessage,
