+++ src/execution/gateway.ts
@@ -1,15 +1,16 @@
 /**
- * Spark Bridge SDK Helpers
+ * Spark Gateway SDK helpers.
  *
- * Calldata encoding for SparkBridge contract operations (withdrawals,
- * token resolution) and high-level withdraw functions using viem.
+ * Calldata encoding for SparkGateway contract operations (withdrawals,
+ * Spark-token address resolution) and high-level withdraw helpers using
+ * viem.
  *
  * @example
  * ```typescript
- * import { encodeWithdrawSats, queryBridgedTokenAddress } from "@flashnet/sdk";
+ * import { encodeWithdrawSats, querySparkTokenAddress } from "@flashnet/sdk";
  *
  * const calldata = encodeWithdrawSats(sparkRecipientHex);
- * const tokenAddr = await queryBridgedTokenAddress(rpcUrl, bridgeAddress, sparkTokenIdHex);
+ * const tokenAddr = await querySparkTokenAddress(rpcUrl, gatewayAddress, sparkTokenIdHex);
  * ```
  */
 
@@ -18,13 +19,13 @@ import {
   zeroAddress,
   type Address,
 } from "viem";
-import { sparkBridgeAbi } from "./abis/sparkBridge";
+import { sparkGatewayAbi } from "./abis/sparkGateway";
 import { getClient } from "./rpc";
 
 // Calldata encoding
 
 /**
- * Encode calldata for SparkBridge.withdrawSats(sparkRecipient).
+ * Encode calldata for SparkGateway.withdrawSats(sparkRecipient).
  *
  * @param sparkRecipient - 33-byte compressed public key as hex (with or without 0x prefix)
  * @returns 0x-prefixed calldata hex
@@ -34,14 +35,14 @@ export function encodeWithdrawSats(sparkRecipient: string): string {
     ? sparkRecipient
     : `0x${sparkRecipient}`;
   return encodeFunctionData({
-    abi: sparkBridgeAbi,
+    abi: sparkGatewayAbi,
     functionName: "withdrawSats",
     args: [hex as `0x${string}`],
   });
 }
 
 /**
- * Encode calldata for SparkBridge.withdrawBtkn(tokenAddress, amount, sparkRecipient).
+ * Encode calldata for SparkGateway.withdrawBtkn(tokenAddress, amount, sparkRecipient).
  */
 export function encodeWithdrawToken(
   tokenAddress: string,
@@ -52,21 +53,21 @@ export function encodeWithdrawToken(
     ? sparkRecipient
     : `0x${sparkRecipient}`;
   return encodeFunctionData({
-    abi: sparkBridgeAbi,
+    abi: sparkGatewayAbi,
     functionName: "withdrawBtkn",
     args: [tokenAddress as Address, amount, hex as `0x${string}`],
   });
 }
 
-// Bridge queries
+// Gateway queries
 
 /**
- * Query the EVM token address for a bridged Spark token.
- * Returns null if not yet bridged.
+ * Query the EVM ERC20 address for a Spark token deployed by the gateway.
+ * Returns null if no token has been deployed yet for that Spark token id.
  */
-export async function queryBridgedTokenAddress(
+export async function querySparkTokenAddress(
   rpcUrl: string,
-  bridgeAddress: string,
+  gatewayAddress: string,
   sparkTokenIdHex: string
 ): Promise<string | null> {
   const client = getClient(rpcUrl);
@@ -75,8 +76,8 @@ export async function queryBridgedTokenAddress(
     : `0x${sparkTokenIdHex}`;
 
   const result = await client.readContract({
-    address: bridgeAddress as Address,
-    abi: sparkBridgeAbi,
+    address: gatewayAddress as Address,
+    abi: sparkGatewayAbi,
     functionName: "tokenBySparkId",
     args: [hex as `0x${string}`],
   });
@@ -86,20 +87,20 @@ export async function queryBridgedTokenAddress(
 }
 
 /**
- * Poll for a bridged token address until it appears or timeout.
+ * Poll for a Spark token's EVM address until it appears or timeout.
  */
-export async function waitForBridgedTokenAddress(
+export async function waitForSparkTokenAddress(
   rpcUrl: string,
-  bridgeAddress: string,
+  gatewayAddress: string,
   sparkTokenIdHex: string,
   timeoutMs: number = 20_000,
   pollIntervalMs: number = 500
 ): Promise<string | null> {
   const deadline = Date.now() + timeoutMs;
   while (Date.now() < deadline) {
-    const addr = await queryBridgedTokenAddress(
+    const addr = await querySparkTokenAddress(
       rpcUrl,
-      bridgeAddress,
+      gatewayAddress,
       sparkTokenIdHex
     );
     if (addr) return addr;
@@ -107,4 +108,3 @@ export async function waitForBridgedTokenAddress(
   }
   return null;
 }
-
